[FEAT] ImagePickerViewController의 싱글톤 사용을 위한 SignUpViewController extension 삭제 및 UIViewController extension 파일 추가 #16
